Update render-list to use class
I also change its name because I will create an index.js for render-list module
